refactor - LottoController 클래스의 일부 로직을 해당 도메인으로 이동
