+++ src/main/java/domain/Lotto.java
@@ -1,5 +1,7 @@
 package domain;
 
+import java.util.List;
+
 public class Lotto {
     private LottoNumbers lottoNumbers;
 
@@ -9,6 +11,15 @@ public class Lotto {
 
     public static Lotto createLotto() {
         return new Lotto(new LottoNumbers().shuffleNumbers());
+    }
+
+    public int matchLottoNumbers(LottoNumbers winningLotto) {
+        List<Integer> winningNumbers = winningLotto.getNumbers();
+
+        return (int) lottoNumbers.getNumbers()
+                .stream()
+                .filter(winningNumbers::contains)
+                .count();
     }
 
     public LottoNumbers getLottoNumbers() {
